Remove GetListMixin
